requirements.txt
numpy 
pandas 
matplotlib 
seaborn 
scikit-learn 
pillow 
jupyter